requirements.txt
autopep8==1.6.0
et-xmlfile==1.1.0
lxml==4.7.1
openpyxl==3.0.9
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-docx==0.8.11
python-dotenv==0.19.2
toml==0.10.2
XlsxWriter==3.0.2
